requirements.txt
streamlit
crewai
autogen
Langchain
langchain-community
python-dotenv
googlesearch-python
openai
requests
chromadb
nltk
